requirements.txt
PyQt5>=5.15.7
Pillow>=9.3.0